add first edition

+++ tohtml.sh
@@ -3,9 +3,12 @@
 
 for file in $(find . -name *.ipynb)
 do
-   jupyter nbconvert "$file" --to html
    htmlFile="${file/.ipynb/.html}"
    dest="./papers/html-nbconv/${htmlFile:2}"
+   [ ! -f "$dest" ] || continue
+
+   jupyter nbconvert "$file" --to html
+
    mkdir -p `dirname "$dest"`
    mv -v "$htmlFile" "$dest"
    # cp -Rv "$htmlFile" "$dest"
